decouple GrantCLRCalculation and move to own file
